CurrencyRestController, CurrencyMapper, JoinWalletMapper and changeCu…

+++ src/main/java/com/telerikacademy/web/virtualwallet/services/CurrencyServiceImpl.java
@@ -1,8 +1,13 @@
 package com.telerikacademy.web.virtualwallet.services;
 
+import com.telerikacademy.web.virtualwallet.exceptions.AuthorizationException;
+import com.telerikacademy.web.virtualwallet.exceptions.EntityDuplicateException;
+import com.telerikacademy.web.virtualwallet.exceptions.EntityNotFoundException;
 import com.telerikacademy.web.virtualwallet.models.Currency;
+import com.telerikacademy.web.virtualwallet.models.User;
 import com.telerikacademy.web.virtualwallet.repositories.contracts.CurrencyRepository;
 import com.telerikacademy.web.virtualwallet.services.contracts.CurrencyService;
+import com.telerikacademy.web.virtualwallet.utils.UserRole;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,6 +15,7 @@ import java.util.List;
 @Service
 public class CurrencyServiceImpl implements CurrencyService {
 
+    public static final String ONLY_ADMINS_ERR = "Only admins can manage currencies";
     private final CurrencyRepository currencyRepository;
 
     @Autowired
@@ -27,20 +33,52 @@ public class CurrencyServiceImpl implements CurrencyService {
     public List<Currency> getAll() {
         return currencyRepository.getAll();
     }
-
     @Override
-    public void create(Currency currency) {
+    public void create(Currency currency,User user) {
+        checkIsDuplicated(currency);
+        checkAdmin(user);
         currencyRepository.create(currency);
     }
 
     @Override
-    public void update(Currency currency) {
+    public void update(Currency currency,User user) {
+        checkIsDuplicated(currency, currency.getId());
+        checkAdmin(user);
         currencyRepository.update(currency);
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id,User user) {
+        checkAdmin(user);
         currencyRepository.delete(id);
+    }
+
+    private void checkAdmin(User user){
+        if (!isAdmin(user)){
+            throw new AuthorizationException(ONLY_ADMINS_ERR);
+        }
+    }
+
+    private boolean isAdmin(User user) {
+        return user.getUserRoles().stream().anyMatch(r -> r.getRoleType().equals(UserRole.admin.toString()));
+    }
+
+    private void checkIsDuplicated(Currency currency) {
+        try {
+            currencyRepository.getByField("currencyCode", currency.getCurrencyCode());
+            throw new EntityDuplicateException("Currency","code", currency.getCurrencyCode());
+        } catch (EntityNotFoundException ignore) {
+        }
+    }
+
+    private void checkIsDuplicated(Currency currency, int id) {
+        try {
+            Currency currencyToCheck = currencyRepository.getByField("currencyCode", currency.getCurrencyCode());
+            if (currencyToCheck.getId() != id) {
+                throw new EntityDuplicateException("Currency", "code", currency.getCurrencyCode());
+            }
+        } catch (EntityNotFoundException ignore) {
+        }
     }
